+++ packages/transport-grpc/src/send-execute-script.js
@@ -2,18 +2,18 @@ import {invariant} from "@onflow/util-invariant"
 import {ExecuteScriptAtLatestBlockRequest, ExecuteScriptAtBlockIDRequest, ExecuteScriptAtBlockHeightRequest, AccessAPI} from "@onflow/protobuf"
 import {unary as defaultUnary} from "./unary"
 
-const argumentBuffer = arg => Buffer.from(JSON.stringify(arg), "utf8")
-const hexBuffer = hex => Buffer.from(hex, "hex")
+const argumentBuffer = (arg, context) => context.Buffer.from(JSON.stringify(arg), "utf8")
+const hexBuffer = (hex, context) => context.Buffer.from(hex, "hex")
 
 async function sendExecuteScriptAtBlockIDRequest(ix, context, opts) {
   const unary = opts.unary || defaultUnary
 
   const req = new ExecuteScriptAtBlockIDRequest()
 
-  req.setBlockId(hexBuffer(ix.block.id))
+  req.setBlockId(hexBuffer(ix.block.id, context))
 
-  const code = Buffer.from(ix.message.cadence, "utf8")
-  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument)))
+  const code = context.Buffer.from(ix.message.cadence, "utf8")
+  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument, context)))
   req.setScript(code)
 
   const res = await unary(opts.node, AccessAPI.ExecuteScriptAtBlockID, req, context)
@@ -28,8 +28,8 @@ async function sendExecuteScriptAtBlockHeightRequest(ix, context, opts) {
 
   req.setBlockHeight(Number(ix.block.height))
 
-  const code = Buffer.from(ix.message.cadence, "utf8")
-  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument)))
+  const code = context.Buffer.from(ix.message.cadence, "utf8")
+  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument, context)))
   req.setScript(code)
 
   const res = await unary(opts.node, AccessAPI.ExecuteScriptAtBlockHeight, req, context) 
@@ -42,8 +42,8 @@ async function sendExecuteScriptAtLatestBlockRequest(ix, context, opts) {
 
   const req = new ExecuteScriptAtLatestBlockRequest()
   
-  const code = Buffer.from(ix.message.cadence, "utf8")
-  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument)))
+  const code = context.Buffer.from(ix.message.cadence, "utf8")
+  ix.message.arguments.forEach(arg => req.addArguments(argumentBuffer(ix.arguments[arg].asArgument, context)))
   req.setScript(code)
 
   const res = await unary(opts.node, AccessAPI.ExecuteScriptAtLatestBlock, req, context)
@@ -54,7 +54,7 @@ async function sendExecuteScriptAtLatestBlockRequest(ix, context, opts) {
 function constructResponse(ix, context, res)  {
   let ret = context.response()
   ret.tag = ix.tag
-  ret.encodedData = JSON.parse(Buffer.from(res.getValue_asU8()).toString("utf8"))
+  ret.encodedData = JSON.parse(context.Buffer.from(res.getValue_asU8()).toString("utf8"))
 
   return ret
 }
@@ -62,6 +62,7 @@ function constructResponse(ix, context, res)  {
 export async function sendExecuteScript(ix, context = {}, opts = {}) {
   invariant(opts.node, `SDK Send Execute Script Error: opts.node must be defined.`)
   invariant(context.response, `SDK Send Execute Script Error: context.response must be defined.`)
+  invariant(context.Buffer, `SDK Send Execute Script Error: context.Buffer must be defined.`)
   
   ix = await ix